refactor(server): address ORM v2 parity review feedback and split the query builder spec
Skip alias reference quoting inside SQL string literals, reject a join
alias that collides with the main query alias, gate explicit-condition
to-many joins to raw reads like implicit ones, base relation null
hydration on the joined id instead of an all-null heuristic so partial
selections keep their object, make select('<mainAlias>') override a
prior findOptions selection, and drop redundant inline comments.

Split the query builder spec into per-concern files under 400 lines
with a shared table-shape fixture module.
