--- packages/twenty-server/src/engine/twenty-orm-v2/query-builder/__tests__/workspace-select-query-builder-v2.spec.ts
@@ -1,1500 +1,0 @@
-import { FieldMetadataType } from 'twenty-shared/types';
-import { And, Equal, In, LessThan, Not } from 'typeorm';
-
-import { RelationType } from 'src/engine/metadata-modules/field-metadata/interfaces/relation-type.interface';
-import { type CompiledStatement } from 'src/engine/twenty-orm-v2/sql/utils/compile-named-parameters.util';
-import { TwentyOrmV2Exception } from 'src/engine/twenty-orm-v2/exceptions/twenty-orm-v2.exception';
-import { WorkspaceSelectQueryBuilderV2 } from 'src/engine/twenty-orm-v2/query-builder/workspace-select-query-builder-v2';
-import { buildColumnResultAlias } from 'src/engine/twenty-orm-v2/sql/utils/build-column-result-alias.util';
-import { type WorkspaceTableShape } from 'src/engine/twenty-orm-v2/table-shape/types/workspace-table-shape.type';
-
-const SCHEMA_NAME = 'workspace_1wgvd1injqtife6y4rvfbu3h5';
-
-const buildColumn = (
-  columnName: string,
-  compositeParentFieldName?: string,
-) => ({
-  columnName,
-  fieldMetadataId: `field-${columnName}`,
-  fieldName: compositeParentFieldName ?? columnName,
-  fieldMetadataType: FieldMetadataType.TEXT,
-  ...(compositeParentFieldName !== undefined
-    ? { compositeParentFieldName }
-    : {}),
-});
-
-const companyTableShape: WorkspaceTableShape = {
-  objectMetadataId: 'company-object-id',
-  nameSingular: 'company',
-  schemaName: SCHEMA_NAME,
-  tableName: 'company',
-  columnShapeByColumnName: {
-    id: buildColumn('id'),
-    name: buildColumn('name'),
-    deletedAt: buildColumn('deletedAt'),
-  },
-  columnNames: ['id', 'name', 'deletedAt'],
-  relationShapeByFieldName: {
-    person: {
-      fieldName: 'person',
-      fieldMetadataId: 'field-company',
-      relationType: RelationType.MANY_TO_ONE,
-      targetObjectMetadataId: 'person-object-id',
-      targetFieldMetadataId: 'field-people',
-      joinColumnName: 'personId',
-    },
-  },
-  hasDeletedAtColumn: true,
-};
-
-const personTableShape: WorkspaceTableShape = {
-  objectMetadataId: 'person-object-id',
-  nameSingular: 'person',
-  schemaName: SCHEMA_NAME,
-  tableName: 'person',
-  columnShapeByColumnName: {
-    id: buildColumn('id'),
-    nameFirstName: buildColumn('nameFirstName', 'name'),
-    nameLastName: buildColumn('nameLastName', 'name'),
-    companyId: buildColumn('companyId'),
-    deletedAt: buildColumn('deletedAt'),
-  },
-  columnNames: [
-    'id',
-    'nameFirstName',
-    'nameLastName',
-    'companyId',
-    'deletedAt',
-  ],
-  relationShapeByFieldName: {
-    company: {
-      fieldName: 'company',
-      fieldMetadataId: 'field-company',
-      relationType: RelationType.MANY_TO_ONE,
-      targetObjectMetadataId: 'company-object-id',
-      targetFieldMetadataId: 'field-people',
-      joinColumnName: 'companyId',
-    },
-    people: {
-      fieldName: 'people',
-      fieldMetadataId: 'field-people',
-      relationType: RelationType.ONE_TO_MANY,
-      targetObjectMetadataId: 'company-object-id',
-      targetFieldMetadataId: 'field-company',
-    },
-  },
-  hasDeletedAtColumn: true,
-};
-
-const buildQueryBuilder = ({
-  rows = [],
-  tableShape = personTableShape,
-}: {
-  rows?: Record<string, unknown>[];
-  tableShape?: WorkspaceTableShape;
-} = {}) => {
-  const executedStatements: CompiledStatement[] = [];
-
-  const queryBuilder = new WorkspaceSelectQueryBuilderV2('person', {
-    tableShape,
-    executor: {
-      execute: async (statement) => {
-        executedStatements.push(statement);
-
-        return rows;
-      },
-    },
-    objectRecordsPermissions: {},
-    tableShapeByObjectMetadataId: () => companyTableShape,
-    onBeforeExecute: () => undefined,
-    formatResult: (records) => records as never,
-  });
-
-  return { queryBuilder, executedStatements };
-};
-
-describe('WorkspaceSelectQueryBuilderV2', () => {
-  it('should select the requested columns with alias-prefixed output names', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({
-      select: { id: true, nameFirstName: true },
-    });
-
-    expect(queryBuilder.getQuery()).toBe(
-      'SELECT "person"."id" AS "person_id", "person"."nameFirstName" AS "person_nameFirstName" ' +
-        `FROM "${SCHEMA_NAME}"."person" AS "person" ` +
-        'WHERE "person"."deletedAt" IS NULL',
-    );
-  });
-
-  it('should compile named parameters coming from the shared filter parser', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder
-      .setFindOptions({ select: { id: true } })
-      .andWhere('"person"."nameFirstName" ILIKE :name7f3a', {
-        name7f3a: '%ac%',
-      });
-
-    const [text, values] = queryBuilder.getQueryAndParameters();
-
-    expect(text).toContain('"person"."nameFirstName" ILIKE $1');
-    expect(values).toEqual(['%ac%']);
-  });
-
-  it('should add the soft-delete predicate unless withDeleted was called', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    expect(queryBuilder.getQuery()).toContain('"person"."deletedAt" IS NULL');
-
-    queryBuilder.withDeleted();
-    expect(queryBuilder.getQuery()).not.toContain(
-      '"person"."deletedAt" IS NULL',
-    );
-  });
-
-  it('should join a to-one relation on its foreign key', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-
-    expect(queryBuilder.getQuery()).toContain(
-      `LEFT JOIN "${SCHEMA_NAME}"."company" AS "company" ` +
-        'ON ("person"."companyId" = "company"."id")',
-    );
-  });
-
-  it('should render an inner join for a to-one relation', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.innerJoin('person.company', 'company');
-
-    expect(queryBuilder.getQuery()).toContain(
-      `INNER JOIN "${SCHEMA_NAME}"."company" AS "company" ` +
-        'ON ("person"."companyId" = "company"."id")',
-    );
-  });
-
-  it('should resolve a join path rooted on a joined alias', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.leftJoin('company.person', 'companyPerson');
-
-    expect(queryBuilder.getQuery()).toContain(
-      `LEFT JOIN "${SCHEMA_NAME}"."company" AS "companyPerson" ` +
-        'ON ("company"."personId" = "companyPerson"."id")',
-    );
-  });
-
-  it('should reject a join path rooted on an unknown alias', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    expect(() => queryBuilder.leftJoin('unknown.company', 'company')).toThrow(
-      TwentyOrmV2Exception,
-    );
-  });
-
-  it('should report a to-many join to the shared guard rather than rendering it', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.people', 'people');
-
-    expect(queryBuilder.expressionMap.joinAttributes).toEqual([
-      {
-        alias: { name: 'people' },
-        relation: { isOneToMany: true, isManyToMany: false },
-      },
-    ]);
-    expect(() => queryBuilder.getQuery()).toThrow(TwentyOrmV2Exception);
-  });
-
-  it('should project every joined column for a leftJoinAndSelect', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoinAndSelect('person.company', 'company');
-
-    const sql = queryBuilder.getQuery();
-
-    expect(sql).toContain(
-      '"company"."id" AS "company_id", "company"."name" AS "company_name", ' +
-        '"company"."deletedAt" AS "company_deletedAt"',
-    );
-    expect(sql).toContain(
-      `LEFT JOIN "${SCHEMA_NAME}"."company" AS "company" ` +
-        'ON ("person"."companyId" = "company"."id")',
-    );
-  });
-
-  it('should render an inner join for an innerJoinAndSelect', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.innerJoinAndSelect('person.company', 'company');
-
-    expect(queryBuilder.getQuery()).toContain(
-      `INNER JOIN "${SCHEMA_NAME}"."company" AS "company"`,
-    );
-  });
-
-  it('should hydrate joined columns under the relation property', async () => {
-    const { queryBuilder } = buildQueryBuilder({
-      rows: [
-        {
-          person_id: 'person-1',
-          company_id: 'company-1',
-          company_name: 'Acme',
-          company_deletedAt: null,
-        },
-      ],
-    });
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoinAndSelect('person.company', 'company');
-
-    await expect(queryBuilder.getMany()).resolves.toEqual([
-      {
-        id: 'person-1',
-        company: { id: 'company-1', name: 'Acme', deletedAt: null },
-      },
-    ]);
-  });
-
-  it('should hydrate an unmatched left-joined relation as null', async () => {
-    const { queryBuilder } = buildQueryBuilder({
-      rows: [
-        {
-          person_id: 'person-1',
-          company_id: null,
-          company_name: null,
-          company_deletedAt: null,
-        },
-      ],
-    });
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoinAndSelect('person.company', 'company');
-
-    await expect(queryBuilder.getMany()).resolves.toEqual([
-      { id: 'person-1', company: null },
-    ]);
-  });
-
-  it('should hydrate a partial relation column selected without an output alias', async () => {
-    const { queryBuilder } = buildQueryBuilder({
-      rows: [{ person_id: 'person-1', company_name: 'Acme' }],
-    });
-
-    // The selection may precede the join that introduces the alias.
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.addSelect('company.name');
-    queryBuilder.leftJoin('person.company', 'company');
-
-    expect(queryBuilder.getQuery()).toContain(
-      '"company"."name" AS "company_name"',
-    );
-    await expect(queryBuilder.getMany()).resolves.toEqual([
-      { id: 'person-1', company: { name: 'Acme' } },
-    ]);
-  });
-
-  it('should keep an aliased relation column selection as raw output', async () => {
-    const { queryBuilder } = buildQueryBuilder({
-      rows: [{ person_id: 'person-1', companyName: 'Acme' }],
-    });
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.addSelect('company.name', 'companyName');
-
-    expect(queryBuilder.getQuery()).toContain(
-      '"company"."name" AS "companyName"',
-    );
-    await expect(queryBuilder.getMany()).resolves.toEqual([{ id: 'person-1' }]);
-  });
-
-  it('should reject a partial selection on an unknown alias or column', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-
-    queryBuilder.addSelect('unknown.name');
-    expect(() => queryBuilder.getQuery()).toThrow(TwentyOrmV2Exception);
-  });
-
-  it('should treat selecting the main alias as selecting every main column', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.select('person');
-
-    const sql = queryBuilder.getQuery();
-
-    expect(sql).toContain('"person"."id" AS "person_id"');
-    expect(sql).toContain('"person"."nameFirstName" AS "person_nameFirstName"');
-  });
-
-  it('should report joined projections and every joined alias for permissions', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoinAndSelect('person.company', 'company');
-    queryBuilder.leftJoin('person.people', 'people');
-
-    const columnNamesByAlias = queryBuilder.getReferencedColumnNamesByAlias();
-
-    expect(columnNamesByAlias['company']).toEqual(['id', 'name', 'deletedAt']);
-    expect(columnNamesByAlias['people']).toEqual([]);
-  });
-
-  it('should still reject a to-many joinAndSelect for an entity-hydrating read', async () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoinAndSelect('person.people', 'people');
-
-    await expect(queryBuilder.getMany()).rejects.toThrow(TwentyOrmV2Exception);
-  });
-
-  it('should render an added join condition on a selected inner join', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.innerJoinAndSelect('person.company', 'company');
-    queryBuilder.addJoinCondition('company', '"company"."name" = :rlsName');
-    queryBuilder.setParameters({ rlsName: 'Acme' });
-
-    const sql = queryBuilder.getQuery();
-
-    expect(sql).toContain('("company"."name" = :rlsName)');
-    expect(sql.indexOf('"company"."name" = :rlsName')).toBeLessThan(
-      sql.indexOf('WHERE'),
-    );
-  });
-
-  it('should render a plain to-many join for a raw read', async () => {
-    const { queryBuilder, executedStatements } = buildQueryBuilder();
-
-    queryBuilder.select([]);
-    queryBuilder.addSelect('person.id', 'id');
-    queryBuilder.innerJoin('person.people', 'people');
-
-    await queryBuilder.getRawMany();
-
-    expect(executedStatements[0].text).toContain(
-      `INNER JOIN "${SCHEMA_NAME}"."company" AS "people" ` +
-        'ON ("people"."personId" = "person"."id") AND ("people"."deletedAt" IS NULL)',
-    );
-  });
-
-  it('should render a plain to-many left join for a raw read', async () => {
-    const { queryBuilder, executedStatements } = buildQueryBuilder();
-
-    queryBuilder.select([]);
-    queryBuilder.addSelect('person.id', 'id');
-    queryBuilder.leftJoin('person.people', 'people');
-
-    await queryBuilder.getRawMany();
-
-    expect(executedStatements[0].text).toContain(
-      `LEFT JOIN "${SCHEMA_NAME}"."company" AS "people" ` +
-        'ON ("people"."personId" = "person"."id")',
-    );
-  });
-
-  it('should still reject a plain to-many join for an entity-hydrating read', async () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.people', 'people');
-
-    await expect(queryBuilder.getMany()).rejects.toThrow(TwentyOrmV2Exception);
-    await expect(queryBuilder.getOne()).rejects.toThrow(TwentyOrmV2Exception);
-  });
-
-  it('should count distinct main records when a to-many join multiplies rows', async () => {
-    const { queryBuilder, executedStatements } = buildQueryBuilder({
-      rows: [{ count: '3' }],
-    });
-
-    queryBuilder.innerJoin('person.people', 'people');
-    queryBuilder.where('people.name = :name', { name: 'Acme' });
-    queryBuilder.groupBy('person.id');
-
-    const count = await queryBuilder.getCount();
-
-    expect(count).toBe(3);
-    expect(executedStatements[0].text).toContain(
-      'SELECT COUNT(DISTINCT "person"."id") AS "count"',
-    );
-    expect(executedStatements[0].text).toContain(
-      `INNER JOIN "${SCHEMA_NAME}"."company" AS "people"`,
-    );
-    expect(executedStatements[0].text).not.toContain('GROUP BY');
-  });
-
-  it('should render an added join condition on a plain to-many join', async () => {
-    const { queryBuilder, executedStatements } = buildQueryBuilder();
-
-    queryBuilder.select([]);
-    queryBuilder.addSelect('person.id', 'id');
-    queryBuilder.innerJoin('person.people', 'people');
-    queryBuilder.addJoinCondition('people', '"people"."name" = :ownerName');
-    queryBuilder.setParameters({ ownerName: 'Acme' });
-
-    await queryBuilder.getRawMany();
-
-    expect(executedStatements[0].text).toContain(
-      '("people"."personId" = "person"."id") AND ("people"."name" = $1)',
-    );
-  });
-
-  it('should render a deduped to-many join on the inverse foreign key when the caller opts in', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.people', 'people', undefined, {
-      allowToManyJoin: true,
-    });
-
-    expect(queryBuilder.getQuery()).toContain(
-      `LEFT JOIN (SELECT DISTINCT ON ("personId") * ` +
-        `FROM "${SCHEMA_NAME}"."company" ` +
-        `WHERE "deletedAt" IS NULL ` +
-        `ORDER BY "personId", "id") AS "people" ` +
-        'ON ("people"."personId" = "person"."id")',
-    );
-  });
-
-  it('should emit a plain parameterised LIMIT, with no distinct sub-select', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.limit(31);
-
-    const [text, values] = queryBuilder.getQueryAndParameters();
-
-    expect(text).toContain('LIMIT $');
-    expect(text).not.toContain('DISTINCT');
-    expect(values).toContain(31);
-  });
-
-  it('should reuse one statement shape across page sizes and offsets', () => {
-    const buildPaginatedQuery = (limit: number, offset: number) => {
-      const { queryBuilder } = buildQueryBuilder();
-
-      queryBuilder.setFindOptions({ select: { id: true } });
-      queryBuilder.limit(limit);
-      queryBuilder.offset(offset);
-
-      return queryBuilder.getQueryAndParameters();
-    };
-
-    const [firstText, firstValues] = buildPaginatedQuery(30, 0);
-    const [secondText, secondValues] = buildPaginatedQuery(60, 120);
-
-    expect(firstText).toBe(secondText);
-    expect(firstValues).toEqual([30, 0]);
-    expect(secondValues).toEqual([60, 120]);
-  });
-
-  it('should render order by with direction and nulls handling', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.orderBy({
-      'person.nameFirstName': { order: 'DESC', nulls: 'NULLS LAST' },
-      'company.name': { order: 'ASC' },
-    });
-
-    expect(queryBuilder.getQuery()).toContain(
-      'ORDER BY "person"."nameFirstName" DESC NULLS LAST, "company"."name" ASC',
-    );
-  });
-
-  it('should render a DISTINCT ON clause with normalised columns', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.distinctOn(['company.name', 'nameFirstName']);
-
-    expect(queryBuilder.getQuery()).toContain(
-      'SELECT DISTINCT ON ("company"."name", "person"."nameFirstName") ',
-    );
-  });
-
-  it('should order by an added-select alias without qualifying it', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.select([]);
-    queryBuilder.addSelect('MAX("person"."nameFirstName")', 'max_first_name');
-    queryBuilder.groupBy('person.companyId');
-    queryBuilder.orderBy('max_first_name', 'DESC');
-
-    const sql = queryBuilder.getQuery();
-
-    expect(sql).toContain('ORDER BY "max_first_name" DESC');
-    expect(sql).not.toContain('"person"."max_first_name"');
-  });
-
-  it('should not report an added-select order-by alias as a selected column', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.select([]);
-    queryBuilder.addSelect('MAX("person"."nameFirstName")', 'max_first_name');
-    queryBuilder.orderBy('max_first_name', 'DESC');
-
-    expect(queryBuilder.getReferencedColumnNamesByAlias()['person']).toEqual([
-      'nameFirstName',
-    ]);
-  });
-
-  it('should report distinct-on columns as referenced for permissions', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.distinctOn(['company.name']);
-
-    expect(queryBuilder.getReferencedColumnNamesByAlias()['company']).toEqual([
-      'name',
-    ]);
-  });
-
-  it('should reject a column that does not exist on the object', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { nonExistent: true } });
-
-    expect(() => queryBuilder.getQuery()).toThrow(TwentyOrmV2Exception);
-  });
-
-  it('should nest a bracketed condition built by the shared parsers', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.andWhere({
-      whereFactory: (nested) => {
-        nested.where('"person"."id" = :a', { a: 1 });
-        nested.orWhere('"person"."id" = :b', { b: 2 });
-      },
-    });
-
-    const [text, values] = queryBuilder.getQueryAndParameters();
-
-    expect(text).toContain('(("person"."id" = $1) OR ("person"."id" = $2))');
-    expect(values).toEqual([1, 2]);
-  });
-
-  it('should negate a NotBrackets group instead of rendering it as a plain group', () => {
-    const { queryBuilder } = buildQueryBuilder();
-    const notBrackets = {
-      '@instanceof': Symbol.for('NotBrackets'),
-      whereFactory: (nested: { where: (sql: string) => unknown }) => {
-        nested.where('"person"."nameFirstName" = \'Ada\'');
-      },
-    };
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.andWhere(notBrackets);
-
-    expect(queryBuilder.getQuery()).toContain(
-      'NOT (("person"."nameFirstName" = \'Ada\'))',
-    );
-  });
-
-  it('should render a plain Brackets group without negation', () => {
-    const { queryBuilder } = buildQueryBuilder();
-    const brackets = {
-      '@instanceof': Symbol.for('Brackets'),
-      whereFactory: (nested: { where: (sql: string) => unknown }) => {
-        nested.where('"person"."nameFirstName" = \'Ada\'');
-      },
-    };
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.andWhere(brackets);
-
-    expect(queryBuilder.getQuery()).not.toContain('NOT (');
-  });
-
-  it('should select only aggregate expressions after select([])', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true, nameFirstName: true } });
-    queryBuilder.select([]);
-    queryBuilder.addSelect('COUNT("person"."id")', 'totalCount');
-
-    const sql = queryBuilder.getQuery();
-
-    expect(sql).toContain('SELECT COUNT("person"."id") AS "totalCount" FROM');
-    expect(sql).not.toContain('"person"."nameFirstName"');
-  });
-
-  it('should count columns referenced only inside aggregates as selected', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.select([]);
-    queryBuilder.addSelect('MAX("person"."nameFirstName")', 'maxFirstName');
-
-    expect(queryBuilder.getSelectedColumnNames()).toContain('nameFirstName');
-  });
-
-  it('should keep an explicit empty selection on a clone', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.select([]);
-    queryBuilder.addSelect('COUNT("person"."id")', 'totalCount');
-
-    expect(queryBuilder.clone().getQuery()).not.toContain('"person"."id" AS');
-  });
-
-  it('should carry applied row-level markers onto a clone', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    expect(queryBuilder.markRowLevelPermissionApplied('person')).toBe(true);
-    expect(queryBuilder.clone().markRowLevelPermissionApplied('person')).toBe(
-      false,
-    );
-  });
-
-  it('should clear applied row-level markers when where() replaces the WHERE', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.markRowLevelPermissionApplied('person');
-    queryBuilder.where('"person"."id" = :id', { id: 1 });
-
-    expect(queryBuilder.markRowLevelPermissionApplied('person')).toBe(true);
-  });
-
-  it('should report joined-alias columns referenced only in order by', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.orderBy({ 'company.name': { order: 'ASC' } });
-
-    const columnNamesByAlias = queryBuilder.getReferencedColumnNamesByAlias();
-
-    expect(columnNamesByAlias['person']).toEqual(['id']);
-    expect(columnNamesByAlias['company']).toEqual(['name']);
-  });
-
-  it('should report joined-alias columns referenced only in an added select', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.addSelect('"company"."name"', 'company_name');
-
-    expect(queryBuilder.getReferencedColumnNamesByAlias()['company']).toEqual([
-      'name',
-    ]);
-  });
-
-  it('should quote bare alias.column references in raw where SQL', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.where('company.name = :name', { name: 'Acme' });
-    queryBuilder.andWhere('person.id IN (:...ids)', { ids: ['a'] });
-
-    const sql = queryBuilder.getQuery();
-
-    expect(sql).toContain('("company"."name" = :name)');
-    expect(sql).toContain('("person"."id" IN (:...ids))');
-  });
-
-  it('should quote bare alias.column references inside select and order expressions', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.select([]);
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.addSelect('MAX(company.name)', 'max_name');
-    queryBuilder.groupBy('person.id');
-    queryBuilder.addOrderBy('MIN(company.name)', 'ASC');
-
-    const sql = queryBuilder.getQuery();
-
-    expect(sql).toContain('MAX("company"."name") AS "max_name"');
-    expect(sql).toContain('ORDER BY MIN("company"."name") ASC');
-  });
-
-  it('should attribute aggregate references over a joined alias to that alias', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.select([]);
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.addSelect('MAX(company.name)', 'max_name');
-
-    expect(queryBuilder.getReferencedColumnNamesByAlias()['company']).toEqual([
-      'name',
-    ]);
-  });
-
-  it('should keep the soft-delete predicate outside an OR chain', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.where('"person"."id" = :a', { a: 1 });
-    queryBuilder.orWhere('"person"."id" = :b', { b: 2 });
-
-    expect(queryBuilder.getQuery()).toContain(
-      'WHERE (("person"."id" = :a) OR ("person"."id" = :b)) AND "person"."deletedAt" IS NULL',
-    );
-  });
-
-  it('should filter soft-deleted joined rows in the ON clause', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-
-    const sql = queryBuilder.getQuery();
-
-    expect(sql).toContain('("company"."deletedAt" IS NULL)');
-    expect(sql.indexOf('"company"."deletedAt"')).toBeLessThan(
-      sql.indexOf('WHERE'),
-    );
-  });
-
-  it('should add a joined predicate to ON rather than WHERE', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.addJoinCondition('company', '"company"."name" = :owner');
-    queryBuilder.setParameters({ owner: 'Acme' });
-
-    const sql = queryBuilder.getQuery();
-
-    expect(sql).toContain('("company"."name" = :owner)');
-    expect(sql.indexOf('"company"."name"')).toBeLessThan(sql.indexOf('WHERE'));
-  });
-
-  it('should restore the previous limit when getOne fails', async () => {
-    const failingBuilder = new WorkspaceSelectQueryBuilderV2('person', {
-      tableShape: personTableShape,
-      executor: {
-        execute: async () => {
-          throw new Error('connection lost');
-        },
-      },
-      objectRecordsPermissions: {},
-      tableShapeByObjectMetadataId: () => companyTableShape,
-      onBeforeExecute: () => undefined,
-      formatResult: (records) => records as never,
-    });
-
-    failingBuilder.setFindOptions({ select: { id: true } });
-    failingBuilder.limit(25);
-
-    await expect(failingBuilder.getOne()).rejects.toThrow('connection lost');
-    expect(failingBuilder.getQueryAndParameters()[1]).toEqual([25]);
-  });
-
-  it('should not attribute a joined column to the main alias', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.addSelect('"company"."name"', 'company_name');
-
-    const columnNamesByAlias = queryBuilder.getReferencedColumnNamesByAlias();
-
-    expect(columnNamesByAlias['person']).toEqual(['id']);
-    expect(columnNamesByAlias['company']).toEqual(['name']);
-  });
-
-  it('should keep joined row-level markers when where() replaces the WHERE', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.leftJoin('person.company', 'company');
-    queryBuilder.markRowLevelPermissionApplied('person');
-    queryBuilder.markRowLevelPermissionApplied('company');
-    queryBuilder.where('"person"."id" = :id', { id: 1 });
-
-    expect(queryBuilder.markRowLevelPermissionApplied('person')).toBe(true);
-    expect(queryBuilder.markRowLevelPermissionApplied('company')).toBe(false);
-  });
-
-  it('should strip the alias prefix when mapping rows back to records', async () => {
-    const { queryBuilder } = buildQueryBuilder({
-      rows: [
-        {
-          person_id: 'person-1',
-          person_nameFirstName: 'Ada',
-          company_name: 'Acme',
-        },
-      ],
-    });
-
-    queryBuilder.setFindOptions({ select: { id: true, nameFirstName: true } });
-
-    const records = await queryBuilder.getMany();
-
-    expect(records).toEqual([{ id: 'person-1', nameFirstName: 'Ada' }]);
-  });
-
-  it('should not widen the projection when a relation is joined', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.leftJoin('person.company', 'company');
-
-    expect(queryBuilder.getQuery()).toContain(
-      'SELECT "person"."id" AS "person_id" FROM',
-    );
-  });
-
-  it('should keep result aliases within the postgres identifier limit', async () => {
-    const longColumnName = `veryLongSubfield${'X'.repeat(60)}`;
-    const otherLongColumnName = `veryLongSubfield${'Y'.repeat(60)}`;
-    const resultAlias = buildColumnResultAlias('person', longColumnName);
-    const otherResultAlias = buildColumnResultAlias(
-      'person',
-      otherLongColumnName,
-    );
-
-    const { queryBuilder } = buildQueryBuilder({
-      tableShape: {
-        ...personTableShape,
-        columnShapeByColumnName: {
-          ...personTableShape.columnShapeByColumnName,
-          [longColumnName]: buildColumn(longColumnName),
-          [otherLongColumnName]: buildColumn(otherLongColumnName),
-        },
-      },
-      rows: [{ [resultAlias]: 'left', [otherResultAlias]: 'right' }],
-    });
-
-    queryBuilder.setFindOptions({
-      select: { [longColumnName]: true, [otherLongColumnName]: true },
-    });
-
-    expect(resultAlias.length).toBeLessThanOrEqual(63);
-    expect(resultAlias).not.toBe(otherResultAlias);
-    expect(queryBuilder.getQuery()).toContain(
-      `"person"."${longColumnName}" AS "${resultAlias}"`,
-    );
-    await expect(queryBuilder.getMany()).resolves.toEqual([
-      { [longColumnName]: 'left', [otherLongColumnName]: 'right' },
-    ]);
-  });
-
-  it('should refuse a caller parameter that collides with a reserved name', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    expect(() =>
-      queryBuilder.andWhere('"person"."id" = :ormV2Limit', { ormV2Limit: 1 }),
-    ).toThrow(TwentyOrmV2Exception);
-  });
-
-  it('should report the columns it selected so permissions do not have to parse SQL', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true, companyId: true } });
-
-    expect(queryBuilder.getSelectedColumnNames()).toEqual(['id', 'companyId']);
-  });
-
-  it('should keep a clone independent of the builder it came from', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-
-    const clone = queryBuilder.clone();
-
-    clone.andWhere('"person"."id" = :id', { id: 1 });
-
-    expect(queryBuilder.getQuery()).not.toContain('"person"."id" = :id');
-    expect(clone.getQuery()).toContain('"person"."id" = :id');
-  });
-
-  it('should send one parameterised statement to the executor', async () => {
-    const { queryBuilder, executedStatements } = buildQueryBuilder();
-
-    queryBuilder
-      .setFindOptions({ select: { id: true } })
-      .andWhere('"person"."nameFirstName" = :name', { name: 'Ada' });
-
-    await queryBuilder.getMany();
-
-    expect(executedStatements).toHaveLength(1);
-    expect(executedStatements[0].text).toContain('$1');
-    expect(executedStatements[0].values).toEqual(['Ada']);
-  });
-
-  it('should map take and skip to LIMIT and OFFSET parameters', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder
-      .setFindOptions({ select: { id: true } })
-      .take(31)
-      .skip(60);
-
-    const [text, values] = queryBuilder.getQueryAndParameters();
-
-    expect(text).toContain('LIMIT $1');
-    expect(text).toContain('OFFSET $2');
-    expect(values).toEqual([31, 60]);
-  });
-
-  it('should build a COUNT statement that ignores projection, order and pagination', async () => {
-    const { queryBuilder, executedStatements } = buildQueryBuilder({
-      rows: [{ count: '7' }],
-    });
-
-    queryBuilder
-      .setFindOptions({ select: { id: true } })
-      .orderBy('"person"."id"', 'ASC')
-      .take(10);
-
-    const count = await queryBuilder.getCount();
-
-    expect(count).toBe(7);
-    expect(executedStatements[0].text).toBe(
-      'SELECT COUNT(1) AS "count" ' +
-        `FROM "${SCHEMA_NAME}"."person" AS "person" ` +
-        'WHERE "person"."deletedAt" IS NULL',
-    );
-  });
-
-  it('should render an object-literal where with In as a bound IN clause', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder
-      .setFindOptions({ select: { id: true } })
-      .where({ id: In(['a', 'b']) });
-
-    const [text, values] = queryBuilder.getQueryAndParameters();
-
-    expect(text).toContain('"person"."id" IN ($1, $2)');
-    expect(values).toEqual(['a', 'b']);
-  });
-
-  it('should allocate unique parameter names across object-literal where clauses', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder
-      .setFindOptions({ select: { id: true } })
-      .where({ id: In(['a']) })
-      .andWhere({ companyId: In(['b']) });
-
-    const [, values] = queryBuilder.getQueryAndParameters();
-
-    expect(values).toEqual(['a', 'b']);
-  });
-
-  it('should reject an object-literal where on an unknown column', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    expect(() => queryBuilder.where({ missing: In(['x']) })).toThrow(
-      TwentyOrmV2Exception,
-    );
-  });
-
-  it('should render comparison operators in an object-literal where', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.where({ nameFirstName: LessThan('x') });
-
-    const [text, values] = queryBuilder.getQueryAndParameters();
-
-    expect(text).toContain('"person"."nameFirstName" < $1');
-    expect(values).toContain('x');
-  });
-
-  it('should combine operators with And', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.where({ nameFirstName: And(LessThan('z'), Not(Equal('a'))) });
-
-    const [text] = queryBuilder.getQueryAndParameters();
-
-    expect(text).toContain(
-      '("person"."nameFirstName" < $1 AND NOT ("person"."nameFirstName" = $2))',
-    );
-  });
-
-  it('should negate a nested operator with Not', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder.setFindOptions({ select: { id: true } });
-    queryBuilder.where({ id: Not(In(['a', 'b'])) });
-
-    const [text] = queryBuilder.getQueryAndParameters();
-
-    expect(text).toContain('NOT ("person"."id" IN ($1, $2))');
-  });
-
-  it('should treat a plain value in an object-literal where as equality', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder
-      .setFindOptions({ select: { id: true } })
-      .where({ companyId: 'company-1' });
-
-    const [text, values] = queryBuilder.getQueryAndParameters();
-
-    expect(text).toContain('("person"."companyId" = $1)');
-    expect(values).toEqual(['company-1']);
-  });
-
-  it('should treat the Equal operator in an object-literal where as equality', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder
-      .setFindOptions({ select: { id: true } })
-      .where({ companyId: Equal('company-1') });
-
-    const [text, values] = queryBuilder.getQueryAndParameters();
-
-    expect(text).toContain('("person"."companyId" = $1)');
-    expect(values).toEqual(['company-1']);
-  });
-
-  it('should treat a null value in an object-literal where as IS NULL', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder
-      .setFindOptions({ select: { id: true } })
-      .where({ companyId: null });
-
-    expect(queryBuilder.getQuery()).toContain('("person"."companyId" IS NULL)');
-  });
-
-  it('should emit a GROUP BY with aggregate and grouped columns', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder
-      .select([])
-      .addSelect('COUNT(*)', 'totalCount')
-      .addSelect('"person"."companyId"', 'companyId')
-      .groupBy('"person"."companyId"');
-
-    expect(queryBuilder.getQuery()).toBe(
-      'SELECT COUNT(*) AS "totalCount", "person"."companyId" AS "companyId" ' +
-        `FROM "${SCHEMA_NAME}"."person" AS "person" ` +
-        'WHERE "person"."deletedAt" IS NULL ' +
-        'GROUP BY "person"."companyId"',
-    );
-  });
-
-  it('should append additional GROUP BY expressions with addGroupBy', () => {
-    const { queryBuilder } = buildQueryBuilder();
-
-    queryBuilder
-      .select([])
-      .addSelect('COUNT(*)', 'totalCount')
-      .groupBy('"person"."companyId"')
-      .addGroupBy('"person"."nameFirstName"');
-
-    expect(queryBuilder.getQuery()).toContain(
-      'GROUP BY "person"."companyId", "person"."nameFirstName"',
-    );
-  });
-
-  it('should return every raw row from getRawMany', async () => {
-    const { queryBuilder, executedStatements } = buildQueryBuilder({
-      rows: [
-        { companyId: 'c1', totalCount: '3' },
-        { companyId: 'c2', totalCount: '1' },
-      ],
-    });
-
-    queryBuilder
-      .select([])
-      .addSelect('COUNT(*)', 'totalCount')
-      .addSelect('"person"."companyId"', 'companyId')
-      .groupBy('"person"."companyId"');
-
-    const rows = await queryBuilder.getRawMany();
-
-    expect(rows).toEqual([
-      { companyId: 'c1', totalCount: '3' },
-      { companyId: 'c2', totalCount: '1' },
-    ]);
-    expect(executedStatements[0].text).toContain(
-      'GROUP BY "person"."companyId"',
-    );
-  });
-});
-
-// The timeline messaging service is the reference consumer of the TypeORM parity
-// surface: these specs mirror its three query shapes end to end.
-describe('WorkspaceSelectQueryBuilderV2 timeline messaging parity', () => {
-  const buildShape = (
-    objectMetadataId: string,
-    nameSingular: string,
-    columnNames: string[],
-    relationShapeByFieldName: WorkspaceTableShape['relationShapeByFieldName'],
-  ): WorkspaceTableShape => ({
-    objectMetadataId,
-    nameSingular,
-    schemaName: SCHEMA_NAME,
-    tableName: nameSingular,
-    columnShapeByColumnName: Object.fromEntries(
-      columnNames.map((columnName) => [columnName, buildColumn(columnName)]),
-    ),
-    columnNames,
-    relationShapeByFieldName,
-    hasDeletedAtColumn: true,
-  });
-
-  const messageThreadShape = buildShape(
-    'thread-id',
-    'messageThread',
-    ['id', 'deletedAt'],
-    {
-      messages: {
-        fieldName: 'messages',
-        fieldMetadataId: 'field-thread-messages',
-        relationType: RelationType.ONE_TO_MANY,
-        targetObjectMetadataId: 'message-id',
-        targetFieldMetadataId: 'field-message-thread',
-      },
-    },
-  );
-
-  const messageShape = buildShape(
-    'message-id',
-    'message',
-    ['id', 'messageThreadId', 'receivedAt', 'deletedAt'],
-    {
-      messageThread: {
-        fieldName: 'messageThread',
-        fieldMetadataId: 'field-message-thread',
-        relationType: RelationType.MANY_TO_ONE,
-        targetObjectMetadataId: 'thread-id',
-        targetFieldMetadataId: 'field-thread-messages',
-        joinColumnName: 'messageThreadId',
-      },
-      messageParticipants: {
-        fieldName: 'messageParticipants',
-        fieldMetadataId: 'field-message-participants',
-        relationType: RelationType.ONE_TO_MANY,
-        targetObjectMetadataId: 'participant-id',
-        targetFieldMetadataId: 'field-participant-message',
-      },
-      messageChannelMessageAssociations: {
-        fieldName: 'messageChannelMessageAssociations',
-        fieldMetadataId: 'field-message-associations',
-        relationType: RelationType.ONE_TO_MANY,
-        targetObjectMetadataId: 'association-id',
-        targetFieldMetadataId: 'field-association-message',
-      },
-    },
-  );
-
-  const messageParticipantShape = buildShape(
-    'participant-id',
-    'messageParticipant',
-    [
-      'id',
-      'messageId',
-      'personId',
-      'workspaceMemberId',
-      'role',
-      'handle',
-      'deletedAt',
-    ],
-    {
-      message: {
-        fieldName: 'message',
-        fieldMetadataId: 'field-participant-message',
-        relationType: RelationType.MANY_TO_ONE,
-        targetObjectMetadataId: 'message-id',
-        targetFieldMetadataId: 'field-message-participants',
-        joinColumnName: 'messageId',
-      },
-      person: {
-        fieldName: 'person',
-        fieldMetadataId: 'field-participant-person',
-        relationType: RelationType.MANY_TO_ONE,
-        targetObjectMetadataId: 'timeline-person-id',
-        targetFieldMetadataId: null,
-        joinColumnName: 'personId',
-      },
-      workspaceMember: {
-        fieldName: 'workspaceMember',
-        fieldMetadataId: 'field-participant-member',
-        relationType: RelationType.MANY_TO_ONE,
-        targetObjectMetadataId: 'member-id',
-        targetFieldMetadataId: null,
-        joinColumnName: 'workspaceMemberId',
-      },
-    },
-  );
-
-  const timelinePersonShape = buildShape(
-    'timeline-person-id',
-    'person',
-    ['id', 'nameFirstName', 'avatarUrl', 'deletedAt'],
-    {},
-  );
-
-  const workspaceMemberShape = buildShape(
-    'member-id',
-    'workspaceMember',
-    ['id', 'nameFirstName', 'avatarUrl', 'deletedAt'],
-    {},
-  );
-
-  const associationShape = buildShape(
-    'association-id',
-    'messageChannelMessageAssociation',
-    ['id', 'messageId', 'messageChannelId', 'deletedAt'],
-    {
-      message: {
-        fieldName: 'message',
-        fieldMetadataId: 'field-association-message',
-        relationType: RelationType.MANY_TO_ONE,
-        targetObjectMetadataId: 'message-id',
-        targetFieldMetadataId: 'field-message-associations',
-        joinColumnName: 'messageId',
-      },
-    },
-  );
-
-  const shapeByObjectMetadataId: Record<string, WorkspaceTableShape> = {
-    'thread-id': messageThreadShape,
-    'message-id': messageShape,
-    'participant-id': messageParticipantShape,
-    'timeline-person-id': timelinePersonShape,
-    'member-id': workspaceMemberShape,
-    'association-id': associationShape,
-  };
-
-  const buildTimelineQueryBuilder = (
-    alias: string,
-    tableShape: WorkspaceTableShape,
-    rows: Record<string, unknown>[] = [],
-  ) => {
-    const executedStatements: CompiledStatement[] = [];
-
-    const queryBuilder = new WorkspaceSelectQueryBuilderV2(alias, {
-      tableShape,
-      executor: {
-        execute: async (statement) => {
-          executedStatements.push(statement);
-
-          return rows;
-        },
-      },
-      objectRecordsPermissions: {},
-      tableShapeByObjectMetadataId: (objectMetadataId) =>
-        shapeByObjectMetadataId[objectMetadataId],
-      onBeforeExecute: () => undefined,
-      formatResult: (records) => records as never,
-    });
-
-    return { queryBuilder, executedStatements };
-  };
-
-  it('should count threads across two to-many inner joins', async () => {
-    const { queryBuilder, executedStatements } = buildTimelineQueryBuilder(
-      'messageThread',
-      messageThreadShape,
-      [{ count: '2' }],
-    );
-
-    const count = await queryBuilder
-      .innerJoin('messageThread.messages', 'messages')
-      .innerJoin('messages.messageParticipants', 'messageParticipants')
-      .where('messageParticipants.personId IN(:...personIds)', {
-        personIds: ['person-1'],
-      })
-      .groupBy('messageThread.id')
-      .getCount();
-
-    expect(count).toBe(2);
-
-    const sql = executedStatements[0].text;
-
-    expect(sql).toContain('SELECT COUNT(DISTINCT "messageThread"."id")');
-    expect(sql).toContain(
-      `INNER JOIN "${SCHEMA_NAME}"."message" AS "messages" ` +
-        'ON ("messages"."messageThreadId" = "messageThread"."id")',
-    );
-    expect(sql).toContain(
-      `INNER JOIN "${SCHEMA_NAME}"."messageParticipant" AS "messageParticipants" ` +
-        'ON ("messageParticipants"."messageId" = "messages"."id")',
-    );
-    expect(sql).toContain('"messageParticipants"."personId" IN($1)');
-  });
-
-  it('should page thread ids ordered by an aggregate select alias', async () => {
-    const { queryBuilder, executedStatements } = buildTimelineQueryBuilder(
-      'messageThread',
-      messageThreadShape,
-      [{ id: 'thread-1', max_received_at: new Date('2026-01-01') }],
-    );
-
-    const rows = await queryBuilder
-      .select('messageThread.id', 'id')
-      .addSelect('MAX(messages.receivedAt)', 'max_received_at')
-      .innerJoin('messageThread.messages', 'messages')
-      .innerJoin('messages.messageParticipants', 'messageParticipants')
-      .where('messageParticipants.personId IN (:...personIds)', {
-        personIds: ['person-1'],
-      })
-      .groupBy('messageThread.id')
-      .orderBy('max_received_at', 'DESC')
-      .offset(3)
-      .limit(6)
-      .getRawMany();
-
-    expect(rows[0]).toEqual({
-      id: 'thread-1',
-      max_received_at: new Date('2026-01-01'),
-    });
-
-    const sql = executedStatements[0].text;
-
-    expect(sql).toContain(
-      'SELECT "messageThread"."id" AS "id", ' +
-        'MAX("messages"."receivedAt") AS "max_received_at"',
-    );
-    expect(sql).toContain('GROUP BY "messageThread"."id"');
-    expect(sql).toContain('ORDER BY "max_received_at" DESC');
-    expect(sql).toContain('LIMIT $');
-    expect(sql).toContain('OFFSET $');
-  });
-
-  it('should load distinct thread participants with hydrated relations', async () => {
-    const receivedAt = new Date('2026-02-02');
-    const { queryBuilder, executedStatements } = buildTimelineQueryBuilder(
-      'messageParticipant',
-      messageParticipantShape,
-      [
-        {
-          messageParticipant_id: 'participant-1',
-          messageParticipant_messageId: 'message-1',
-          messageParticipant_personId: 'person-1',
-          messageParticipant_workspaceMemberId: null,
-          messageParticipant_role: 'from',
-          messageParticipant_handle: 'ada@acme.dev',
-          messageParticipant_deletedAt: null,
-          person_id: 'person-1',
-          person_nameFirstName: 'Ada',
-          person_avatarUrl: '',
-          person_deletedAt: null,
-          workspaceMember_id: null,
-          workspaceMember_nameFirstName: null,
-          workspaceMember_avatarUrl: null,
-          workspaceMember_deletedAt: null,
-          message_messageThreadId: 'thread-1',
-          message_receivedAt: receivedAt,
-        },
-      ],
-    );
-
-    const participants = await queryBuilder
-      .select('messageParticipant')
-      .addSelect('message.messageThreadId')
-      .addSelect('message.receivedAt')
-      .leftJoinAndSelect('messageParticipant.person', 'person')
-      .leftJoinAndSelect(
-        'messageParticipant.workspaceMember',
-        'workspaceMember',
-      )
-      .leftJoin('messageParticipant.message', 'message')
-      .where('message.messageThreadId = ANY(:messageThreadIds)', {
-        messageThreadIds: ['thread-1'],
-      })
-      .andWhere('messageParticipant.role = :role', { role: 'from' })
-      .orderBy('message.messageThreadId')
-      .distinctOn(['message.messageThreadId', 'messageParticipant.handle'])
-      .getMany();
-
-    expect(participants).toEqual([
-      {
-        id: 'participant-1',
-        messageId: 'message-1',
-        personId: 'person-1',
-        workspaceMemberId: null,
-        role: 'from',
-        handle: 'ada@acme.dev',
-        deletedAt: null,
-        person: {
-          id: 'person-1',
-          nameFirstName: 'Ada',
-          avatarUrl: '',
-          deletedAt: null,
-        },
-        workspaceMember: null,
-        message: { messageThreadId: 'thread-1', receivedAt },
-      },
-    ]);
-
-    const sql = executedStatements[0].text;
-
-    expect(sql).toContain(
-      'SELECT DISTINCT ON ("message"."messageThreadId", "messageParticipant"."handle")',
-    );
-    expect(sql).toContain(
-      `LEFT JOIN "${SCHEMA_NAME}"."person" AS "person" ` +
-        'ON ("messageParticipant"."personId" = "person"."id")',
-    );
-    expect(sql).toContain('"message"."messageThreadId" = ANY($1)');
-    expect(sql).toContain('("messageParticipant"."role" = $2)');
-    expect(sql).toContain('ORDER BY "message"."messageThreadId" ASC');
-  });
-
-  it('should map thread visibility rows through raw to-many left joins', async () => {
-    const { queryBuilder, executedStatements } = buildTimelineQueryBuilder(
-      'messageThread',
-      messageThreadShape,
-      [{ id: 'thread-1', messageChannelId: 'channel-1' }],
-    );
-
-    const rows = await queryBuilder
-      .select('messageThread.id', 'id')
-      .addSelect(
-        'messageChannelMessageAssociation.messageChannelId',
-        'messageChannelId',
-      )
-      .leftJoin('messageThread.messages', 'message')
-      .leftJoin(
-        'message.messageChannelMessageAssociations',
-        'messageChannelMessageAssociation',
-      )
-      .where('messageThread.id = ANY(:messageThreadIds)', {
-        messageThreadIds: ['thread-1'],
-      })
-      .getRawMany();
-
-    expect(rows).toEqual([{ id: 'thread-1', messageChannelId: 'channel-1' }]);
-
-    const sql = executedStatements[0].text;
-
-    expect(sql).toContain(
-      'SELECT "messageThread"."id" AS "id", ' +
-        '"messageChannelMessageAssociation"."messageChannelId" AS "messageChannelId"',
-    );
-    expect(sql).toContain(
-      `LEFT JOIN "${SCHEMA_NAME}"."message" AS "message" ` +
-        'ON ("message"."messageThreadId" = "messageThread"."id")',
-    );
-    expect(sql).toContain(
-      `LEFT JOIN "${SCHEMA_NAME}"."messageChannelMessageAssociation" ` +
-        'AS "messageChannelMessageAssociation" ' +
-        'ON ("messageChannelMessageAssociation"."messageId" = "message"."id")',
-    );
-    expect(sql).toContain('"messageThread"."id" = ANY($1)');
-  });
-});
